fix: dbusmenu related symbols are hidden in Qt6
Dbusmenu related symbols are hidden in Qt6 now, we just can't partially
customize it. Either fully override it (Do not depend any symbol in QtGui)
or just leave it default. Cause we just make a little customization,
just leave it default in Qt6 and try to contribute our improvement
to upstream.

Related-Commit: 78fa103aadf39f19b3f613647a7d3931b4b27e57
Log: fix dbusmenu related symbols are hidden in Qt6